+++ .idea/gradle.xml
@@ -14,6 +14,13 @@
             <option value="$PROJECT_DIR$/material-1.2.2" />
           </set>
         </option>
+        <option name="myModules">
+          <set>
+            <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/material-1.2.2" />
+          </set>
+        </option>
       </GradleProjectSettings>
     </option>
   </component>
