+++ app/src/main/AndroidManifest.xml
@@ -19,10 +19,10 @@
         </activity>
         <activity
             android:name=".StartUp"
-            android:label="@string/app_name"></activity>
+            android:label="@string/app_name" />
         <activity
-            android:name=".PersonalDataForm"
-            android:label="@string/title_activity_personal_data_form"
+            android:name=".ScrollingActivity"
+            android:label="@string/title_activity_scrolling"
             android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
