+++ app/src/main/res/layout/start_up.xml
@@ -27,7 +27,6 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:inputType="textPersonName"
-                    android:text="Name"
                     android:ems="10"
                     android:id="@+id/editText"
                     android:layout_marginTop="29dp"
